Added read only mode, in which node creating, delting and connecting is forbiden

+++ Editor/Scripts/Behaviour Tree/BehaviourTreeView.cs
@@ -23,6 +23,11 @@ namespace i5.VirtualAgents.Editor.BehaviourTrees
         public new class UxmlFactory : UxmlFactory<BehaviourTreeView, UxmlTraits> { }
         public BehaviorTreeAsset Tree;
 
+        /// <summary>
+        /// If yes, the view forbits node creating, delting, connecting, and moving. The individual data from nodes (e.g. the target from a MovementTask) can however still be altered.
+        /// </summary>
+        public bool ReadOnly = false;
+
         public BehaviourTreeView()
         {
             Insert(0, new GridBackground());
@@ -76,26 +81,34 @@ namespace i5.VirtualAgents.Editor.BehaviourTrees
 
         private GraphViewChange OnGraphViewChanged(GraphViewChange graphViewChange)
         {
-            //Delete the removed nodes from the asset
-            if (graphViewChange.elementsToRemove != null)
+            if (!ReadOnly)
             {
-                foreach (var elemToRemove in graphViewChange.elementsToRemove)
+                //Delete the removed nodes from the asset
+                if (graphViewChange.elementsToRemove != null)
                 {
-                    NodeView nodeToRemove = elemToRemove as NodeView;
-                    if (nodeToRemove != null)
+                    foreach (var elemToRemove in graphViewChange.elementsToRemove)
                     {
-                        Tree.DeleteNode(nodeToRemove.node);
+                        NodeView nodeToRemove = elemToRemove as NodeView;
+                        if (nodeToRemove != null)
+                        {
+                            Tree.DeleteNode(nodeToRemove.node);
+                        }
+                    }
+                }
+
+                //Add new edges to the asset
+                if (graphViewChange.edgesToCreate != null)
+                {
+                    foreach (var edge in graphViewChange.edgesToCreate)
+                    {
+                        ((NodeView)edge.output.node).node.Children.Add(((NodeView)edge.input.node).node);
                     }
                 }
             }
-
-            //Add new edges to the asset
-            if(graphViewChange.edgesToCreate != null)
+            else
             {
-                foreach (var edge in graphViewChange.edgesToCreate)
-                {
-                    ((NodeView)edge.output.node).node.Children.Add(((NodeView)edge.input.node).node);
-                }
+                graphViewChange.elementsToRemove?.Clear();
+                graphViewChange.edgesToCreate?.Clear();
             }
 
             return graphViewChange;
@@ -108,6 +121,7 @@ namespace i5.VirtualAgents.Editor.BehaviourTrees
         /// <param name="evt"></param>
         public override void BuildContextualMenu(ContextualMenuPopulateEvent evt)
         {
+       
             void CreateVisualNode(ISerializable node, Vector2 position)
             {
                 VisualNode visualNode = Tree.AddNode(node);
@@ -135,9 +149,12 @@ namespace i5.VirtualAgents.Editor.BehaviourTrees
                 }
             }
 
-            BuildContextMenuEntrysFromType<IAgentTask>("Tasks");
-            BuildContextMenuEntrysFromType<ICompositeNode>("Composite Nodes");
-            BuildContextMenuEntrysFromType<IDecoratorNode>("Decorator Nodes");
+            if (!ReadOnly)
+            {
+                BuildContextMenuEntrysFromType<IAgentTask>("Tasks");
+                BuildContextMenuEntrysFromType<ICompositeNode>("Composite Nodes");
+                BuildContextMenuEntrysFromType<IDecoratorNode>("Decorator Nodes");
+            }
         }
 
 
@@ -161,6 +178,6 @@ namespace i5.VirtualAgents.Editor.BehaviourTrees
             return ports.ToList().Where(endPort =>
             endPort.direction != startPort.direction &&
             endPort.node != startPort.node).ToList();
-        }
+        }     
     }
 }
